Fixes issue where attachment points are unclickable on a large screen. On large displays attachment points would disappear when approached with the mouse.
Note: This uncovered a secondary issue, for molecules with a large
number of inputs there is overlap in the "hit box" for different
attachment points which can cause ambiguity in which attachment point
will receive a connection when the mouseclick is released. This issue
pre-existed these changes, I'll attempt to address it in a follow-up
commit.